ai_agent/requirements.txt
# AI Agent
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0

# Google Gemini SDK (official, replaces deprecated google-generativeai)
google-genai>=1.0.0
groq>=0.4.0

# HTTP client for Django API calls
httpx==0.28.1

# Retry logic
tenacity==9.0.0

# Environment
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
